+++ app/views/Book/Books.php
@@ -1,7 +1,7 @@
-<div>
+<div class="container mt-5">
     <h1>All Books</h1>
     <!-- Table to display all books -->
-    <table border="1">
+    <table class="table table-bordered  ">
         <thead>
             <!-- Table headers -->
             <th>ID</th>
@@ -9,7 +9,11 @@
             <th>Title</th>
             <th>Author</th>
             <th>Date Added</th>
-            <th colspan="3">Actions</th>
+            <?php if ($isAuthenticated): ?>
+                <th colspan="3">Actions</th>
+            <?php else: ?>
+                <th>Action</th>
+            <?php endif; ?>
         </thead>
         <!-- Loop through each book in the data array -->
         <?php foreach ($data['books'] as $book) : ?>
@@ -21,12 +25,16 @@
                 <td><?= $book['author'] ?></td>
                 <td><?= $book['date_added'] ?></td>
                 <!-- Action links for each book -->
-                <td><a href="<?=BASE_URL?>book/delete/<?= $book['id']; ?>">Delete</a></td>
-                <td><a href="<?=BASE_URL?>book/update/<?= $book['id']; ?>">Update</a></td>
+                <?php if ($isAuthenticated): ?>
+                    <td><a href="<?= BASE_URL ?>book/delete/<?= $book['id']; ?>">Delete</a></td>
+                    <td><a href="<?= BASE_URL ?>book/update/<?= $book['id']; ?>">Update</a></td>
+                <?php endif; ?>
                 <td><a href="<?=BASE_URL?>book/id/<?= $book['id']; ?>">View</a></td>
             </tr>
         <?php endforeach; ?>
     </table>
     <!-- Link to add a new book -->
-    <a href="<?=BASE_URL?>book/add">Add new book</a>
+    <?php if ($isAuthenticated): ?>
+        <a href="<?= BASE_URL ?>book/add" class="btn btn-primary">Add new book</a>
+    <?php endif; ?>
 </div>
